Add betting and credit system with Double Down support
Players now start with an initial credit and can place bets before each round. Credit is updated based on game outcomes, including special payouts for blackjack and busts. Double Down is now supported as a player action, doubling the bet and limiting the player to one additional card. The UI prompts for bets, displays credit and bet status, and allows Double Down when available. Card counting accuracy is improved with floating-point division for true count calculation. Output is enhanced to keep players informed throughout gameplay.

+++ CardCounter.cs
@@ -87,8 +87,8 @@ namespace _208_Group_Project_Demo
         /// <returns>The current true count as an integer.</returns>
         public double GetTrueCount()
         {
-            int nDecksLeft = game.ShoeCount() / 52;
-            trueCount = (double)runningCount / (double)nDecksLeft;
+            double nDecksLeft = game.ShoeCount() / 52d;
+            trueCount = runningCount / nDecksLeft;
             return trueCount;
         }
